requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.1.1
cycler==0.12.1
diff_gaussian_rasterization @ file:///home/mayu/postg/scg_version0821/submodules/diff-gaussian-rasterization
einops==0.8.1
filelock==3.13.1
fire==0.7.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
hf-xet==1.1.10
huggingface-hub==0.34.5
idna==3.4
imageio==2.35.1
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.9
lazy_loader==0.4
lightning-utilities==0.11.9
lit==15.0.7
loguru==0.7.3
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
networkx==3.0
numpy==1.24.1
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pillow==10.2.0
pillow_heif==0.18.0
plyfile==1.0.3
propcache==0.2.0
pyparsing==3.1.4
pyre-extensions==0.0.29
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.0
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.28.1
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
simple-lama-inpainting==0.1.0
simple_knn @ file:///home/mayu/postg/scg_version0821/submodules/simple-knn
six==1.17.0
sympy==1.13.3
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.19
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchmetrics==1.5.2
torchvision==0.15.1+cu118
tqdm==4.67.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.13
xformers==0.0.19
yacs==0.1.8
yarl==1.15.2
zipp==3.20.2
